pyproject.toml
[project]
name = "dbt-view-importer"
version = "0.1.0"
description = "Import views from BigQuery to dbt"
readme = "README.md"
authors = [{ name = "K-Oxon", email = "ko1011qfp@gmail.com" }]
requires-python = ">=3.9"
dependencies = [
    "click>=8.1.0,<9.0.0",
    "rich>=12.0.0,<13.0.0",
    "google-cloud-bigquery>=3.10.0,<4.0.0",
    "google-cloud-bigquery-datatransfer>=3.11.0,<4.0.0",
    "jinja2>=3.1.2,<4.0.0",
    "pyyaml>=6.0,<7.0",
]

[project.scripts]
bq2dbt = "bq2dbt.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bq2dbt"]

[project.optional-dependencies]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.9.9",
    "pytest>=7.0.0,<8.0.0",
    "pytest-cov>=4.0.0,<5.0.0",
    "pre-commit>=3.0.0,<4.0.0",
]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[dependency-groups]
dev = [
    "pytest>=7.4.4",
]
